LPD-36740 Extracting ServiceTracker from NestedFieldsWriterInterceptor for later reuse
